Huge Update for Spawn and Ghost-Enemy

+++ SourceCode/17520897-17520407/SmallHeart.h
@@ -14,7 +14,7 @@ class SmallHeart :public CItems
 {
 private:
 	DWORD makeTime;
-	int flyDistance;
+	float flyDistance;
 	bool turnAround;
 	int width, height;
 public:
